refactor page imports, prepare for scss
