+++ package.json
@@ -7,6 +7,7 @@
     "@types/react-slick": "^0.23.2",
     "final-form": "^4.10.0",
     "firebase": "^5.5.1",
+    "node-sass": "^4.9.3",
     "react": "^16.5.2",
     "react-dom": "^16.5.2",
     "react-final-form": "^3.6.6",
